refactor: Remove unneeded routes and group by purpose, and Update route paths to include '/v1' versioning
