--- routes/family.js
@@ -1,9 +1,0 @@
-const express = require('express');
-
-const familyController = require('../controllers/family-controller');
-
-const router = express.Router();
-
-router.post('/family', familyController.familyRegistration);
-
-module.exports = router;
