--- routes/forget-password.js
@@ -1,9 +1,0 @@
-const express = require('express');
-
-const userController = require('../controllers/user-controller');
-
-const router = express.Router();
-
-router.post('/forgot-password', userController.requestResetPassword);
-
-module.exports = router;
