--- routes/member.js
@@ -1,9 +1,0 @@
-const express = require('express');
-
-const memberController = require('../controllers/member-controller');
-
-const router = express.Router();
-
-router.post('/member', memberController.memberRegistration);
-
-module.exports = router;
