--- routes/reset-password.js
@@ -1,16 +1,0 @@
-const express = require('express');
-
-const userController = require('../controllers/user-controller');
-
-const router = express.Router();
-
-router.get(
-  '/reset-password/:email/:resetPasswordToken',
-  userController.resetPasswordActivation,
-);
-router.post(
-  '/reset-password/:email/:resetPasswordToken',
-  userController.resetPasswordUpdates,
-);
-
-module.exports = router;
